added currency service

+++ app/src/main/java/com/example/walletlog/User.kt
@@ -2,6 +2,6 @@ package com.example.walletlog
 
 import java.io.Serializable
 
-class User(var id : String, var login : String, var password : String, var budget : Int) : Serializable {
+class User(var id : String, var login : String, var password : String, var budget : Float) : Serializable {
 
 }
